move file outside of controllers folder
