+++ server/app.js
@@ -18,9 +18,9 @@ app.use(cookieParser());
 app.use(express.static(join(__dirname, "public")));
 
 //Mount route handlers
-app.use(require("./routes/controllers/AppointmentAPI"));
-app.use(require("./routes/controllers/eventAPI"));
-app.use(require(".routes/controllers/authentication"));
+app.use(require("./routes/AppointmentAPI"));
+app.use(require("./routes/eventAPI"));
+app.use(require(".routes/authentication"));
 
 //Mount utilities
 app.use(require("./utils/errorHandler"));
